Update and rename README.md to school.grade.calculator
This is a project that i made just for fun, trying to improve the calculator by not accepting number and strings together. I've already done a configuration that doesn't accept stings which show a message "error, try again" but now i have to do the exception for when someone types a string with a number.